python_api/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1
python-dotenv==1.0.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.2
pytz==2023.3
langchain==0.1.1
langchain-openai==0.0.5
langchain-community==0.0.13
openai==1.6.1
pypdf==3.17.1
requests==2.31.0